move websocket mode to seperate package: * should be able to use it in the langserver-go and langserver-antha now. * pending: write tests for the websocket mode.
