+++ langserver/cmd/langserver-antha/langserver-go.go
@@ -7,12 +7,10 @@ import (
 	"io"
 	"log"
 	"net"
-	"net/http"
 	"os"
 
-	"github.com/gorilla/websocket"
-
 	"github.com/sourcegraph/go-langserver/langserver"
+	"github.com/sourcegraph/go-langserver/langserver/modes"
 	"github.com/sourcegraph/jsonrpc2"
 )
 
@@ -21,7 +19,6 @@ var (
 	addr    = flag.String("addr", ":4389", "server listen address (tcp|ws)")
 	trace   = flag.Bool("trace", false, "print all requests and responses")
 	logfile = flag.String("logfile", "", "also log to this file (in addition to stderr)")
-	ctx     = context.Background()
 )
 
 func main() {
@@ -71,21 +68,7 @@ func run() error {
 		}
 
 	case "ws":
-		log.Println("langserver-go: websocket listening on", *addr)
-		http.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
-			log.Printf("langserver-go: wsConn upgrading - w: %p, r: %p", &w, r)
-
-			wsConn, err := upgrader.Upgrade(w, r, nil)
-			if err != nil {
-				log.Printf("langserver-go: wsConn upgrade error - w: %p, r: %p, err: %v", &w, r, err)
-				return
-			}
-
-			log.Printf("langserver-go: wsConn: %p - upgraded - w: %p, r: %p", wsConn, &w, r)
-			wsHandler(w, r, wsConn, connOpt)
-		})
-		err := http.ListenAndServe(*addr, nil)
-		return err
+		return modes.WebSocket(*addr, connOpt);
 
 	case "stdio":
 		log.Println("langserver-go: reading on stdin, writing on stdout")
@@ -96,67 +79,6 @@ func run() error {
 	default:
 		return fmt.Errorf("invalid mode %q", *mode)
 	}
-}
-
-var upgrader = websocket.Upgrader{
-	ReadBufferSize:  1024,
-	WriteBufferSize: 1024,
-	// disable security for now
-	// "If CheckOrigin returns false, you will get the error you described. By default, it returns false if the request is cross-origin."
-	CheckOrigin: func(r *http.Request) bool {
-		return true
-	},
-}
-
-func wsHandler(w http.ResponseWriter, r *http.Request, wsConn *websocket.Conn, connOpt []jsonrpc2.ConnOpt) {
-	// defer wsConn.Close()
-
-	handler := langserver.NewHandler()
-	for {
-		messageType, reader, err := wsConn.NextReader()
-		if err != nil {
-			log.Printf("langserver-go: wsConn: %p - NextReader error - err: %v", wsConn, err)
-			return
-		}
-		log.Printf("langserver-go: wsConn: %p - NextReader - reader: %p, messageType: %d", wsConn, &reader, messageType)
-
-		writer, err := wsConn.NextWriter(messageType)
-		if err != nil {
-			log.Printf("langserver-go: wsConn: %p - NextWriter error - err: %v", wsConn, err)
-			return
-		}
-		log.Printf("langserver-go: wsConn: %p - NextWriter - writer: %p", wsConn, &writer)
-
-		rwc := wsrwc{reader: reader, writer: writer, closer: writer}
-		conn := jsonrpc2.NewConn(ctx, rwc, handler, connOpt...)
-		<-conn.DisconnectNotify()
-		// defer conn.Close()
-		// conn.ReadMessagesStart(ctx, rwc)
-
-	// 	err = conn.Close()
-	// 	if err != nil {
-	// 		log.Printf("langserver-go: wsConn: %p - conn.Close() error - conn: %p, err: %v", wsConn, conn, err)
-	// 		return
-	// 	}
-	}
-}
-
-type wsrwc struct {
-	reader io.Reader
-	writer io.Writer
-	closer io.Closer
-}
-
-func (ws wsrwc) Read(p []byte) (int, error) {
-	return ws.reader.Read(p)
-}
-
-func (ws wsrwc) Write(p []byte) (int, error) {
-	return ws.writer.Write(p)
-}
-
-func (ws wsrwc) Close() error {
-	return ws.closer.Close()
 }
 
 type stdrwc struct{}
